[#925] Migrate to new event concept variant
