+++ iceoryx2-cal/conformance-tests/Cargo.toml
@@ -18,6 +18,7 @@ std = [
   "iceoryx2-bb-concurrency/std",
   "iceoryx2-bb-container/std",
   "iceoryx2-bb-derive-macros/std",
+  "iceoryx2-bb-lock-free/std",
   "iceoryx2-bb-loggers/std",
   "iceoryx2-bb-memory/std",
   "iceoryx2-bb-posix/std",
@@ -39,6 +40,7 @@ iceoryx2-bb-concurrency = { workspace = true }
 iceoryx2-bb-container = { workspace = true }
 iceoryx2-bb-derive-macros = { workspace = true }
 iceoryx2-bb-elementary-traits = { workspace = true }
+iceoryx2-bb-lock-free = { workspace = true }
 iceoryx2-bb-loggers = { workspace = true }
 iceoryx2-bb-memory = { workspace = true }
 iceoryx2-bb-posix = { workspace = true }
