+++ iceoryx2-cal/conformance-tests/src/lib.rs
@@ -21,12 +21,12 @@ pub mod arc_sync_policy_abandonable_trait;
 pub mod arc_sync_policy_trait;
 pub mod communication_channel_trait;
 pub mod dynamic_storage_trait;
-pub mod event_id_tracker_trait;
-pub mod event_signal_mechanism_trait;
+// pub mod event_id_tracker_trait;
+// pub mod event_signal_mechanism_trait;
 pub mod event_trait;
 pub mod monitoring_trait;
-pub mod named_concept_trait;
-pub mod reactor_trait;
+// pub mod named_concept_trait;
+// pub mod reactor_trait;
 pub mod resizable_shared_memory_trait;
 pub mod serialize_trait;
 pub mod shared_memory_trait;
